pyproject.toml
[project]
name = "rf-generators"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Reto Trappitsch", email = "reto@galactic-forensics.space" }
]
dependencies = [
    "pyserial>=3.5",
]
readme = "README.md"
requires-python = ">= 3.11"
license = { text = "MIT" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.3.3",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/rf_generators"]
